split commands into file based APIs and expose natively
