+++ .changeset/funny-ducks-tease.md
@@ -1,0 +1,5 @@
+---
+'modular-scripts': minor
+---
+
+Expose node.js API for modular-scripts
