+++ packages/modular-scripts/src/execSync.ts
@@ -1,0 +1,20 @@
+import execa from 'execa';
+import chalk from 'chalk';
+
+export default function execSync(
+    file: string,
+    args: string[],
+    options: { log?: boolean } & execa.SyncOptions = { log: true },
+  ): execa.ExecaSyncReturnValue<string> {
+    const { log, ...opts } = options;
+    if (log) {
+      console.log(chalk.grey(`$ ${file} ${args.join(' ')}`));
+    }
+    return execa.sync(file, args, {
+      stdin: process.stdin,
+      stderr: process.stderr,
+      stdout: process.stdout,
+      cleanup: true,
+      ...opts,
+    });
+  }
