+++ packages/modular-scripts/src/index.ts
@@ -1,0 +1,4 @@
+export { default as build } from './build';
+export { default as start } from './start';
+export { default as test } from './test';
+export { default as addPackage } from './addPackage';
